Move to minijinja from Tera
I've had some problems with Tera. And minijinja brings support for
embedding templates or rendering fragments (useful for HTMX).

For the next feature the biggest difference is the rendering of float
numbers though. Rendering floats with exactly 1 digit precision is
'involved', whereas in minijinja the round filter just works and doesn't
drop '.0'.