Turn options starting with https:// into links
This is very useful when comparing more complex topics that are
representable as links which can contain more context. For example we
are choosing a Mastodon instance at the moment and are linking to their
respective about pages with their rules.

Security wise this shouldn't be an issue. These links are HTTPS only, no
`javascript:alert(1)` shenanigans, they are opened by the browser, so
they are GET only, not POSTing possible, and they open in a new
tab/window so the chance of masquerading as the same sk-rs instance is
low.