Sort options of closed polls by score
When a poll is closed, people are interested in what the winner was.
This makes it easier for polls with more than 5 options.

Why not sort while the poll is still running? That could change the
order of options while voting, confusing voters. Especially when they
are going down the list from top to bottom they _will_ loose track of
their progress.

@fapdash: This one is for you :-)