Add buttons to close and re-open polls
If the default 90 days is too generous and you want to freeze the
results, this is a quick and simple way to do that. And if you clicked
on the close button out of curiosity or by mistake the re-open button
will quickly "undo" that action.

A more fine-tuned way to change the expiration date would be very
welcome too, but this is rather simple and covers a large portion of
use-cases already.

Closes #13