Fix sorting of closed polls
The option name and resistance score were sorted correctly, but the
table of individual votes was still iterated in the order they were
stored in the database, not according to their score. So the values on
the right hand side of the table would not match the calculated score on
the left hand side, which is very confusing for users.