Fix spacing between table and admin buttons
A simple div with buttons doesn't have margin at the bottom, and a table
(or most other elemts really) doesn't have margin at the top. A section
does. So with a div followed by a table there was no spacing inbetween,
and the table header was butting right up to the admin buttons. Now
there is the same space between the table and the admin buttons as there
is between the buttons and the expiration date above them.